Moved the "GitHubCachedGlobalContributorModel" class into it's own file
